cloud_functions/edinet-document-csv-to-bigquery/requirements.txt
functions-framework==3.5.0
google-cloud-storage
google-cloud-bigquery
google.cloud.logging
pandas
pyarrow